+++ lib/stardag/tests/test__cli/conftest.py
@@ -20,6 +20,7 @@ from stardag.utils.testing.helper_tasks import SyncOnlyTask
 CLI_MODULES = (
     "builds",
     "builds_stop",
+    "builds_frontier",
     "executions",
     "plans",
     "deployments",